Backend/requirements.txt
fastapi
uvicorn
python-multipart
gradio_client
cloudinary
python-dotenv